requirements.txt
django-rest-framework
django-cors-headers
psycopg2-binary
python-dotenv
django-jazzmin
Pillow
django-ckeditor-5
drf-yasg
django-solo
django-phonenumber-field[phonenumbers]